+++ app/src/main/java/com/depromeet/baton/presentation/ui/mypage/view/EditAccountActivity.kt
@@ -18,9 +18,10 @@ import com.depromeet.baton.domain.repository.UserinfoRepository
 import com.depromeet.baton.map.util.NetworkResult
 import com.depromeet.baton.presentation.base.BaseActivity
 import com.depromeet.baton.presentation.base.BaseViewModel
-import com.depromeet.baton.presentation.ui.sign.AddAccountUiState
-import com.depromeet.baton.presentation.ui.sign.AddAccountViewModel
-import com.depromeet.baton.presentation.util.RegexConstant
+import com.depromeet.baton.presentation.bottom.BottomMenuItem
+import com.depromeet.baton.presentation.bottom.BottomSheetFragment
+import com.depromeet.baton.presentation.bottom.BottomSheetFragment.Companion.CHECK_ITEM_VIEW
+import com.depromeet.baton.presentation.ui.mypage.viewmodel.EditAccountViewModel
 import com.depromeet.baton.remote.user.UserAccount
 import com.depromeet.bds.component.BdsToast
 import dagger.hilt.android.AndroidEntryPoint
@@ -37,6 +38,11 @@ class EditAccountActivity :BaseActivity<ActivityEditAccountBinding>(R.layout.act
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setObserver()
+        setListener()
+    }
+
+    private fun setListener(){
+        binding.appbar.setOnBackwardClick{onBackPressed()}
         binding.appbar.setOnButton{
             viewModel.checkEditOption()
         }
@@ -68,16 +74,25 @@ class EditAccountActivity :BaseActivity<ActivityEditAccountBinding>(R.layout.act
                     }
                     is EditAccountViewModel.ViewEvent.OpenBankSelection -> {
                         //BEAN: 은행 선택 바텀 시트 추가 필요
-//                    BottomSheetFragment(
-//                        "은행 선택",
-//                        AddAccountViewModel.supportedBanks.map {
-//                            BottomMenuItem(it, false)
-//                        }
-//                    )
+                        showBankBottom()
+
                     }
                 }
                 viewModel.consumeViewEvent(viewEvent)
             }
+    }
+
+    private fun showBankBottom(){
+        val list =  EditAccountViewModel.supportedBanks.map {
+            BottomMenuItem(it, it==viewModel.uiState.value.bank)
+        }
+        val bottom =BottomSheetFragment.newInstance("은행선택",list,CHECK_ITEM_VIEW ,object:
+            BottomSheetFragment.Companion.OnItemClick {
+            override fun onSelectedItem(selected: BottomMenuItem, pos: Int) {
+              viewModel.handleBankSelected(selected.listItem!!)
+            }
+        })
+        bottom.show(supportFragmentManager,null)
     }
 
     companion object{
@@ -87,123 +102,5 @@ class EditAccountActivity :BaseActivity<ActivityEditAccountBinding>(R.layout.act
             intent.putExtra("account",bundle)
             context.startActivity(intent)
         }
-    }
-}
-data class EditAccountUiState(
-    val name: String,
-    val bank: String,
-    val account: String,
-    val checkEdit : Boolean,
-    val onNameChanged: (Editable?) -> Unit,
-    val onBankSelected: (String) -> Unit,
-    val onAccountChanged: (Editable?) -> Unit,
-    val onBankSelectionClick: () -> Unit,
-) {
-    private val isNameValid = name.isNotBlank() && RegexConstant.ONLY_COMPLETE_HANGLE.matches(name)
-
-    //BEAN: 일단 계좌 번호 길이 13으루
-    private val isAccountValid = account.length == 13&&RegexConstant.ONLY_NUMBERS.matches(account)
-
-    val nameErrorReason = if (isNameValid) null else "올바른 예금주명을 입력해주세요."
-    val accountErrorReason = if (isAccountValid) null else "올바른 계좌번호를 입력해주세요."
-
-    val isEnabled =  checkEdit && isNameValid && isAccountValid && bank.isNotBlank()
-    val isCheckedStr = if(checkEdit)"완료" else "수정"
-}
-
-
-@HiltViewModel
-class EditAccountViewModel @Inject constructor(
-    val savedStateHandle: SavedStateHandle,
-    private val accountRepository: AccountRepository
-) : BaseViewModel() {
-
-    private val _uiState: MutableStateFlow<EditAccountUiState> = MutableStateFlow(createState())
-    val uiState = _uiState.asStateFlow()
-
-    private val _viewEvents: MutableStateFlow<List<ViewEvent>> = MutableStateFlow(emptyList())
-    val viewEvents = _viewEvents.asStateFlow()
-
-    private fun createState():EditAccountUiState {
-       val account= savedStateHandle.get<Bundle>("account") as Bundle
-       val info =  account.get("data") as UserAccount
-
-        return EditAccountUiState(
-            name = info.holder,
-            bank = info.bank,
-            account = info.number,
-            checkEdit = false,
-            onNameChanged = ::handleNameChanged,
-            onBankSelected = ::handleBankSelected,
-            onAccountChanged = ::handleAccountChanged,
-            onBankSelectionClick = ::handleBankSelectionClick,
-        )
-    }
-
-    fun editAccount(){
-        viewModelScope.launch {
-            val userId =2
-            runCatching {
-                uiState?.let {
-                    accountRepository.updateAccount(userId,it.value.name, it.value.bank,it.value.account)
-                }
-            }.onSuccess {
-                when(it){
-                    is NetworkResult.Success ->{ addViewEvent(ViewEvent.EditAccountDone)}
-                    is NetworkResult.Error ->{
-                        Timber.e(it.message)
-                        addViewEvent(ViewEvent.EditAccountFailure("계좌변경 실패"))
-                    }
-                }
-            }
-        }
-    }
-
-    fun checkEditOption(){
-        if(uiState.value.isEnabled) editAccount()
-        else if(uiState.value.checkEdit)  addViewEvent(ViewEvent.EditAccountFailure("입력 정보를 확인해주세요."))
-        _uiState.update { it.copy(checkEdit = !it.checkEdit) }
-    }
-
-    private fun handleNameChanged(editable: Editable?) {
-        _uiState.update { it.copy(name = editable.toString()) }
-    }
-
-    fun handleBankSelected(bank: String) {
-        _uiState.update { it.copy(bank = bank) }
-    }
-
-    private fun handleAccountChanged(editable: Editable?) {
-        _uiState.update { it.copy(account = editable.toString()) }
-    }
-
-    private fun handleBankSelectionClick() {
-        val currentBank = uiState.value.bank
-        addViewEvent(ViewEvent.OpenBankSelection(currentBank))
-    }
-
-    private fun addViewEvent(viewEvent: ViewEvent) {
-        _viewEvents.update { it + viewEvent }
-    }
-
-    fun consumeViewEvent(viewEvent: ViewEvent) {
-        _viewEvents.update { it - viewEvent }
-    }
-
-    sealed interface ViewEvent {
-        object EditAccountDone : ViewEvent
-        data class EditAccountFailure(val msg: String) : ViewEvent
-        data class OpenBankSelection(val selectedBank: String) : ViewEvent
-    }
-
-    companion object {
-        val supportedBanks = listOf(
-            "NH농협은행",
-            "KB국민은행",
-            "카카오뱅크",
-            "신한은행",
-            "우리은행",
-            // and more ...
-        )
     }
 }
