+++ app/src/main/res/layout/activity_empty_account.xml
@@ -6,6 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
         <com.depromeet.bds.component.BdsBackwardAppBar
+            android:id="@+id/appbar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:bds_title="계좌 설정"
